Improve coding for UnorderedComparingCollection
