+++ Sources/SafeDICore/Models/TypeDescription.swift
@@ -423,55 +423,6 @@ extension ExprSyntax {
     }
 }
 
-// MARK: - UnorderedComparingCollection
-
-public struct UnorderedComparingCollection<Element: Codable & Hashable & Sendable>: Codable, Hashable, Sendable, Collection, ExpressibleByArrayLiteral {
-
-    // MARK: Initialization
-
-    public init(_ array: [Element]) {
-        self.array = array
-        set = Set(array)
-    }
-
-    // MARK: Equatable
-
-    public static func == (lhs: UnorderedComparingCollection, rhs: UnorderedComparingCollection) -> Bool {
-        lhs.set == rhs.set
-    }
-
-    // MARK: Hashable
-
-    public func hash(into hasher: inout Hasher) {
-        hasher.combine(set)
-    }
-
-    // MARK: Collection
-
-    public func makeIterator() -> IndexingIterator<Array<Element>> { array.makeIterator() }
-    public var startIndex: Int { array.startIndex }
-    public var endIndex: Int { array.endIndex }
-    public func index(after i: Int) -> Int {
-        array.index(after: i)
-    }
-    public subscript(position: Int) -> Element {
-        array[position]
-    }
-
-    // MARK: ExpressibleByArrayLiteral
-
-    public init(arrayLiteral elements: Element...) {
-        self.init(elements)
-    }
-
-    public typealias ArrayLiteralElement = Element
-
-    // MARK: Private
-
-    private let array: [Element]
-    private let set: Set<Element>
-}
-
 // MARK: - GenericArgumentVisitor
 
 private final class GenericArgumentVisitor: SyntaxVisitor {
