+++ Tests/SafeDICoreTests/TypeDescriptionTests.swift
@@ -573,28 +573,6 @@ final class TypeDescriptionTests: XCTestCase {
         )
     }
 
-    func test_UnorderedComparingCollection_makeIterator_iteratesInOrder() {
-        for (index, value) in UnorderedComparingCollection([1, 2, 3]).enumerated() {
-            switch index {
-            case 0:
-                XCTAssertEqual(value, 1)
-            case 1:
-                XCTAssertEqual(value, 2)
-            case 2:
-                XCTAssertEqual(value, 3)
-            case _:
-                XCTFail("Unexpected index \(index)")
-            }
-        }
-    }
-
-    func test_UnorderedComparingCollection_hashesEquivalentCollectionsIdentically() {
-        XCTAssertEqual(
-            UnorderedComparingCollection([1, 2, 3]).hashValue,
-            UnorderedComparingCollection([2, 1, 3]).hashValue
-        )
-    }
-
     // MARK: - Visitors
 
     private final class TypeIdentifierSyntaxVisitor: SyntaxVisitor {
